Updating to include support for VS2012
Now have two projects, one for 2012 (11.0) and one for 2013 (12.0) that
are versioned to match the VS version and the version of the SDK
assemblies that are required to have this function properly.  Nuget
definitions for each as well.  Still waiting on NuGet pull request to be
honoured.  For now simply changing the GUID in your sln file for your
modelproj project to the CSPROJ guid value will allow you to use this
msbuild task inline.